pyproject.toml
[project]
name = "py-pid-analyzer"
version = "2.0.0"
description = "A Python utility for analyzing Linux process IDs with specialized enhanced analysis for different application types"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "psutil>=5.9.0",
]
authors = [
    { name = "PY-PID-ANALYZER" }
]
license = { text = "MIT" }
keywords = ["process", "analysis", "linux", "pid", "systemd", "java", "webserver"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]

[project.optional-dependencies]
aws = ["boto3>=1.26.0"]
yaml = ["pyyaml>=6.0"]
all = ["boto3>=1.26.0", "pyyaml>=6.0"]

[project.scripts]
pid-analyzer = "pid_analyzer.main:main"
analyze = "pid_analyzer.main:main"

[project.urls]
Homepage = "https://github.com/evilerbender/py-pid-analyzer"
Repository = "https://github.com/evilerbender/py-pid-analyzer"
Issues = "https://github.com/evilerbender/py-pid-analyzer/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pid_analyzer"]
